+++ modules/processing_diffusers.py
@@ -651,6 +651,8 @@ def process_diffusers(p: StableDiffusionProcessing, seeds, prompts, negative_pro
             )
             shared.state.sampling_steps = refiner_args['num_inference_steps']
             try:
+                shared.sd_refiner.register_to_config(requires_aesthetics_score=shared.opts.diffusers_aesthetics_score)
+                print('HERE req', shared.sd_refiner.config.requires_aesthetics_score)
                 refiner_output = shared.sd_refiner(**refiner_args) # pylint: disable=not-callable
             except AssertionError as e:
                 shared.log.info(e)